# docs/swagger/swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/EqualNine/Lab01/1.0.0
info:
  version: "1.0.0"
  title: ServerING API
  description: The API for the ServerING project
paths:
...
  '/api/v1/platforms':
    get:
      tags:
          - Platform
      description: Returns All Platforms

      responses:
        '200':
          description: All platforms

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
                
        '204':
            description: No Content
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No Platforms"
    
    post:
      security:
        - bearerAuth: []      

      tags:
        - Platform
      description: Post Platform

      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Authorization Bearer

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: NewPlatform
                cost:
                  type: integer
                  example: 50000
                popularity:
                  type: integer
                  example: 50

      responses:
        '201':
          description: Saved Platform

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
                
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "User must be logged in"
        
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Not enough rights for this"
  '/api/v1/platforms/{id}':
    get:
      tags:
          - Platform
      description: Return Platform by Id

      parameters:
          - in: path
            required: true
            name: id
            schema:
              type: integer
            description: Platform Id

      responses:
        '200':
          description: Platform by Id

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        
        '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No platform with id: '5'"

    put:
        security:
          - bearerAuth: []
          
        tags:
          - Platform

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Platform Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: NewPlatform
                  cost:
                    type: integer
                    example: 50000
                  popularity:
                    type: integer
                    example: 50

        responses:
          '201':
            description: Updated Platform

            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/Platform'        
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No platform with id: '5'"

    patch:
        security:
          - bearerAuth: []
          
        tags:
          - Platform
        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Platform Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: BestPlatform

        responses:
          '201':
            description: Updated Platform
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/Platform'

          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No platform with id: '5'"

    delete:
        security:
          - bearerAuth: []

        tags:
          - Platform
        description: Delete Platform by Id

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
    
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Platform Id
                
        responses:
          '200':
            description: Platform removed
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "Platform deleted successfully!"
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No platform with id: '5'"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # только для документации
  schemas:
...
    Platform:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Play Station 5
        price:
          type: integer
          example: 60000
        popularity:
          type: integer
          example: 100
...